+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.2] - 2024-08-08
+
+### Added
+- Implement `Default` for `GraphView`
+- Implement `Serialize` and `Deserialize` for `GraphView`.
+
+### Changed
+- `GraphView::lowest_common_ancestors` takes a slice of `Borrow<Self>` as argument.
+
+
 ## [0.2.1] - 2024-08-05
 
 ### Fixed